chore(code cleanup): separate component styles into individual stylesheets and scope them to associated component, rename all .js files that have html to .jsx files for clarity
