+++ src/App.css
@@ -26,427 +26,16 @@
 
 body {
   font-size: 16px;
+  width: 100%;
+  min-height: 100vh;
 }
 
 
 
 /* global reusable classes */
-
-/* text control focus style */
 .text-control-focus-style:focus {
   box-shadow: none;
   border: 2px solid var(--text-control-border);
 }
 
 
-.scroll-to-top-btn {
-  background-color:/* #5a50eb**/ white;
-  color: #5a50eb;
-  position: fixed;
-  bottom: 5%;
-  right: 8%;
-  z-index: 10;
-  border: 2px solid #5a50eb;
-  animation: btnMoveUp 2s ease 0s infinite;
-}
-
-.scroll-to-top-btn:hover {
-  background-color: #5a50eb;
-  color: white;
-}
-
-@keyframes btnMoveUp {
-  0% {
-    transform: translateY(10%);
-  }
-
-  20% {
-    transform: translateY(20%);
-  }
-
-  40% {
-    transform: translateY(30%);
-  }
-
-  60% {
-    transform: translateY(40%);
-  }
-
-  /* 80% { transform: translateY(50%);}
-
-    100% { transform: translateY(60%);} */
-}
-
-
-
-
-
-/* start date picker styling */
-.start-date-picker-styling {
-  width: 100%;
-}
-
-/* end date picker styling */
-.end-date-picker-styling {
-  width: 100%;
-}
-
-/* end of date pickers styling */
-
-
-
-
-
-
-
-
-
-
-
-
-
-/* responsive styling */
-@media (max-width: 480px) {
-  /* home page responsive styling */
-  .place-to-stay {
-    margin-top: 30px;
-  }
-
-  .save-big-text {
-    margin-top: 2%;
-    font-size: 1.3rem;
-  }
-
-  .save-big-text-line2 {
-    padding-left: 4%;
-  }
-
-  .get-away-text {
-    font-size: 1.3rem;
-  }
-
-  .card-title {
-    font-size: 1.1rem;
-  }
-
-  .intro-form-catch-phrase {
-    font-size: 1rem;
-  }
-
-  .intro-background-img {
-    width: 100%;
-    /* height: 450px; */
-    margin-bottom: 5%;
-  }
-
-  .get-away-details-text {
-    font-size: 0.87rem;
-    padding-left: 1%;
-    padding-right: 1%;
-  }
-
-  .welcome-phrase {
-    position: relative;
-    bottom: 30rem;
-    z-index: 10;
-    color: white;
-    opacity: 1;
-    padding: 0% 10% 0% 10%;
-    text-align: center;
-    font-family: 'Courgette', cursive;
-  }
-
-  .intro-form {
-    position: relative;
-    bottom: 360px;
-    /* right: 13px; */
-    left: -21px;
-    width: 85%;
-    height: 75%;
-    background-color: white;
-    padding-left: 3%;
-    padding-right: 3%;
-    padding-top: 1%;
-    box-shadow: 0px 1px 3px grey;
-  }
-
-  .world-of-hotels-section {
-    position: relative;
-    bottom: 260px;
-  }
-
-  .plan-staycation-section {
-    margin-top: -60%;
-  }
-
-  .plan-staycation-text {
-    font-size: 1.3rem;
-  }
-
-  .get-best-prices-text {
-    font-size: 1.3rem;
-  }
-
-  .fetching-hotels-text {
-    font-size: 1rem;
-    color: #808080;
-  }
-
-  .view-all-hotels-text-div {
-    cursor: pointer;
-    margin-bottom: 6%;
-  }
-
-  .view-all-hotels-btn {
-    height: 40px;
-  }
-
-  .footer-header {
-    margin-top: 8%;
-  }
-
-  /* end of home page responsive styling */
-
-  /* contact us responsive styling */
-  .contact-us-details-section {
-    width: 100%;
-  }
-
-  /* end of contact us responsive styling */
-
-  /* about us responsive styling */
-  .about-us-main-background {
-    width: 100%;
-    height: 50vh;
-  }
-
-  .img-text {
-    position: relative;
-    padding-left: 8%;
-    color: black;
-    /* color: var(--primary-color); */
-    font-family: 'Lora', serif;
-    font-size: 1.1rem;
-    font-style: normal;
-    font-weight: bold;
-  }
-
-  .about-us-header-section {
-    margin-top: 20%;
-    margin-bottom: 10%;
-    margin-right: 3.1rem;
-  }
-
-  .main-header-text {
-    padding-left: 5%;
-    font-family: 'Lora', serif;
-    color: var(--primary-color);
-    font-weight: bold;
-  }
-
-  .about-us-mission-row {
-    background-image: linear-gradient(to bottom, blue, white);
-    padding: 5%;
-    text-align: center;
-    color: white;
-  }
-
-  .about-us-mission-text {
-    font-size: 2.3rem;
-    font-weight: bold;
-  }
-
-  .team-pic-style {
-    width: 100%;
-  }
-
-  /* end of about us responsive styling */
-
-  /* book hotel responsive styling */
-  .section-sub-header {
-    font-size: 1rem;
-  }
-
-  .booking-hotel-summary-div {
-    border: 1px solid lightgray;
-    /* border-radius: 1%; */
-    padding: 3%;
-  }
-
-  .details-section {
-    border: 1px solid lightgray;
-    border-radius: 2%;
-    padding: 3%;
-  }
-
-  .booking-hotel-extra-details {
-    font-size: 0.84rem;
-    /* font-weight: bold; */
-    padding-bottom: 5%;
-  }
-
-  .booking-hotel-price {
-    font-weight: bold;
-  }
-
-  .form-control-no-text {
-    margin-bottom: 8%;
-  }
-
-  .details-section-sub-div {
-    margin-bottom: 12%;
-  }
-
-  .confirm-booking-details-btn {
-    margin-bottom: 12%;
-  }
-
-  .paystack-btn {
-    margin-bottom: 9%;
-  }
-
-  .book-room-features {
-    background-color: var(--light-blue-color);
-    padding: 7% 14%;
-    text-align: center;
-    border-radius: 5%;
-    /* display: inline-block; */
-  }
-
-  /* end of book hotel responsive styling */
-
-  /* sign up page responsive styling */
-  .sign-up-form {
-    margin-left: 10%;
-    margin-right: 10%;
-    margin-top: 3%;
-  }
-
-  .other-error-text {
-    font-size: 0.78rem;
-    padding-left: 2%;
-    padding-right: 2%;
-  }
-
-  /* end of sign up page responsive styling */
-
-  /* login page responsive styling */
-  .login-form {
-    margin-left: 8%;
-    margin-right: 8%;
-  }
-
-  .login-other-error-text {
-    color: red;
-    font-family: 'Lora', serif;
-    font-size: 0.78rem;
-    padding-left: 2%;
-    padding-right: 2%;
-  }
-
-  /* get room details responsive styling */
-  .selected-room-checkin-dates {
-    margin-bottom: 7%;
-  }
-
-  .start-date-end-date-null-message {
-    margin-left: 1rem;
-    color: red;
-  }
-
-  .selected-room-details-destination-form {
-    padding: 3%;
-  }
-
-  .selected-room-details-destination-column {
-    margin-bottom: 5%;
-  }
-
-  .selected-room-extra-pics-grid {
-    margin-bottom: 5%;
-  }
-
-  .book-now-button-first {
-    width: 87%;
-    margin-bottom: 10%;
-  }
-
-  .book-now-button-last {
-    width: 85%;
-    margin-bottom: 10%;
-  }
-
-  .book-now-btn-row {
-    margin-top: 8%;
-  }
-
-  .hotel-details-review-section {
-    margin-left: 6%;
-    margin-right: 6%;
-    margin-top: 10%;
-  }
-
-  /* end of get room details responsive styling */
-
-  /* get all rooms responsive styling */
-  .fetch-all-hotels-row {
-    margin-bottom: 12%;
-  }
-
-  .go-to-site-button {
-    width: 100%;
-  }
-
-  .fetch-all-hotels-search-tab-section {
-    margin-bottom: 13%;
-    margin-top: 12%;
-    padding: 0 5.5%;
-  }
-
-  .search-hotel-textbox {
-    height: 43px;
-  }
-
-  /* end of fetch all rooms responsive styling */
-
-  /* review section responsive styling */
-  .post-review-section {
-    margin-left: 6%;
-    margin-right: 6%;
-  }
-
-  .selected-room-details-sub-header {
-    margin-bottom: 10%;
-    font-size: 1.2rem;
-    font-weight: bold;
-  }
-
-  .post-review-btn {
-    /* width: 100%; */
-    margin-left: 7%;
-  }
-
-  .reviews-header-row {
-    margin-bottom: 20%;
-  }
-
-  /* end of review section responsive styling */
-
-  /* faq page responsive styling */
-  .faq-question-answer-section {
-    margin-top: 13%;
-  }
-
-  /* end of faq page responsive styling */
-
-  /* google maps styling */
-  .map-parent {
-    width: 90vw;
-    height: 88vh;
-  }
-
-  .map-container {
-    width: 100%;
-    height: 100%;
-  }
-}
