+++ src/Components/DatePickers/endDatePicker.css
@@ -1,0 +1,3 @@
+.end-date-picker-styling {
+  width: 100%;
+}
