+++ src/Components/DatePickers/startDatePicker.css
@@ -1,0 +1,3 @@
+.start-date-picker-styling {
+  width: 100%;
+}
