+++ src/Components/ScrollToTop/scrollToTop.css
@@ -1,0 +1,37 @@
+.scroll-to-top-btn {
+  background-color:/* #5a50eb**/ white;
+  color: #5a50eb;
+  position: fixed;
+  bottom: 5%;
+  right: 8%;
+  z-index: 10;
+  border: 2px solid #5a50eb;
+  animation: btnMoveUp 2s ease 0s infinite;
+}
+
+.scroll-to-top-btn:hover {
+  background-color: #5a50eb;
+  color: white;
+}
+
+@keyframes btnMoveUp {
+  0% {
+    transform: translateY(10%);
+  }
+
+  20% {
+    transform: translateY(20%);
+  }
+
+  40% {
+    transform: translateY(30%);
+  }
+
+  60% {
+    transform: translateY(40%);
+  }
+
+  /* 80% { transform: translateY(50%);}
+
+  100% { transform: translateY(60%);} */
+}
