--- src/Configuration/Maps2.js
@@ -1,38 +1,0 @@
-import { GoogleMap, Marker, useLoadScript } from '@react-google-maps/api';
-import { useMemo } from 'react';
-
-
-
-
-
-
-
-
-const Maps2 = ({ selectedRoomLatitude, selectedRoomLongitude }) => {
-  let map_center = { lat: selectedRoomLatitude, lng: selectedRoomLongitude };
-  const googleMapsAPIKEY = process.env.REACT_APP_GOOGLE_MAPS_API_KEY;
-
-  const { isLoaded } = useLoadScript({
-    googleMapsApiKey: googleMapsAPIKEY,
-  });
-
-  const center = useMemo(() => map_center, []);
-
-  return (
-    <div className="map-parent">
-      {!isLoaded ? (
-        <h1>Loading...</h1>
-      ) : (
-        <GoogleMap
-          mapContainerClassName="map-container"
-          center={center}
-          zoom={17}
-        >
-          <Marker position={map_center} />
-        </GoogleMap>
-      )}
-    </div>
-  );
-};
-
-export default Maps2;
