+++ src/Pages/Home/home.css
@@ -277,3 +277,107 @@
   border-radius: 10%;
 }
 
+
+
+
+
+
+/* responsive styling */
+@media (max-width: 480px) {
+  .place-to-stay {
+    margin-top: 30px;
+  }
+
+  .save-big-text {
+    margin-top: 2%;
+    font-size: 1.3rem;
+  }
+
+  .save-big-text-line2 {
+    padding-left: 4%;
+  }
+
+  .get-away-text {
+    font-size: 1.3rem;
+  }
+
+  .card-title {
+    font-size: 1.1rem;
+  }
+
+  .intro-form-catch-phrase {
+    font-size: 1rem;
+  }
+
+  .intro-background-img {
+    width: 100%;
+    /* height: 450px; */
+    margin-bottom: 5%;
+  }
+
+  .get-away-details-text {
+    font-size: 0.87rem;
+    padding-left: 1%;
+    padding-right: 1%;
+  }
+
+  .welcome-phrase {
+    position: relative;
+    bottom: 30rem;
+    z-index: 10;
+    color: white;
+    opacity: 1;
+    padding: 0% 10% 0% 10%;
+    text-align: center;
+    font-family: 'Courgette', cursive;
+  }
+
+  .intro-form {
+    position: relative;
+    bottom: 360px;
+    /* right: 13px; */
+    left: -21px;
+    width: 85%;
+    height: 75%;
+    background-color: white;
+    padding-left: 3%;
+    padding-right: 3%;
+    padding-top: 1%;
+    box-shadow: 0px 1px 3px grey;
+  }
+
+  .world-of-hotels-section {
+    position: relative;
+    bottom: 260px;
+  }
+
+  .plan-staycation-section {
+    margin-top: -60%;
+  }
+
+  .plan-staycation-text {
+    font-size: 1.3rem;
+  }
+
+  .get-best-prices-text {
+    font-size: 1.3rem;
+  }
+
+  .fetching-hotels-text {
+    font-size: 1rem;
+    color: #808080;
+  }
+
+  .view-all-hotels-text-div {
+    cursor: pointer;
+    margin-bottom: 6%;
+  }
+
+  .view-all-hotels-btn {
+    height: 40px;
+  }
+
+  .footer-header {
+    margin-top: 8%;
+  }
+}
