+++ src/Pages/HotelDetails/getHotelDetails.css
@@ -310,3 +310,71 @@
   width: 100%;
   height: 100%;
 }
+
+
+
+
+
+/* responsive styling */
+@media (max-width: 480px) {
+  .selected-room-checkin-dates {
+    margin-bottom: 7%;
+  }
+
+  .start-date-end-date-null-message {
+    margin-left: 1rem;
+    color: red;
+  }
+
+  .selected-room-details-destination-form {
+    padding: 3%;
+  }
+
+  .selected-room-details-destination-column {
+    margin-bottom: 5%;
+  }
+
+  .selected-room-extra-pics-grid {
+    margin-bottom: 5%;
+  }
+
+  .book-now-button-first {
+    width: 87%;
+    margin-bottom: 10%;
+  }
+
+  .book-now-button-last {
+    width: 85%;
+    margin-bottom: 10%;
+  }
+
+  .book-now-btn-row {
+    margin-top: 8%;
+  }
+
+  .hotel-details-review-section {
+    margin-left: 6%;
+    margin-right: 6%;
+    margin-top: 10%;
+  }
+
+  .post-review-section {
+    margin-left: 6%;
+    margin-right: 6%;
+  }
+
+  .selected-room-details-sub-header {
+    margin-bottom: 10%;
+    font-size: 1.2rem;
+    font-weight: bold;
+  }
+
+  .post-review-btn {
+    margin-left: 7%;
+  }
+
+  .reviews-header-row {
+    margin-bottom: 20%;
+  }
+
+}
